# src/backend/internal/server/v2/docs/openapi.yaml
openapi: '3.0.2'
info:
  title: Спортивные залы
  version: v2
...
paths:
...
  /api/v2/gym/{gymId}:
    get:
      operationId: getGym
      tags:
        - Gym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: "Зал получен"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gym'
        '400':
          description: "Некорректный запрос"  
        '404':
          description: "Зал не найден"  
        '500':
          description: "Внутренняя ошибка сервера"
    put:
      operationId: putGym
      tags:
        - Gym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutGym'
      responses:
        '204':
          description: "Зал обновлен"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Зал не найден"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteGym
      tags:
        - Gym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Зал удален"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Зал не найден"
        '500':
          description: "Внутренняя ошибка сервера"
...
  /api/v2/gym/{gymId}/equipment/{equipmentId}:      
    put:
      operationId: putEquipment
      tags:
        - Equipment
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: equipmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutEquipment'
      responses:
        '204':
          description: "Оборудование обновлено"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Оборудование не найдено"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteEquipment
      tags:
        - Equipment
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: equipmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Оборудование удалено"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Оборудование не найдено"
        '500':
          description: "Внутренняя ошибка сервера"
...
  /api/v2/gym/{gymId}/membership_type/{membershipTypeId}:        
    put:
      operationId: putMembershipType
      tags:
        - MembershipType
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: membershipTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutMembershipType'
      responses:
        '204':
          description: "Абонемент обновлен"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Оборудование не найдено"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteMembershipType
      tags:
        - MembershipType
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: membershipTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Абонемент удален"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Абонемент не найден"
        '500':
          description: "Внутренняя ошибка сервера"
...
  /api/v2/gym/{clientId}/client_membership/{clientMembershipId}:        
    put:
      operationId: putClientMembership
      tags:
        - ClientMembership
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: clientMembershipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutClientMembership'
      responses:
        '204':
          description: "Клиентский абонемент обновлен"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Клиентские абонементы не найдены"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteClientMembership
      tags:
        - ClientMembership
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: clientMembershipId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Клиентский абонмент удалены"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Клиентский абонемент не найден"
        '500':
          description: "Внутренняя ошибка сервера"
...
  /api/v2/payment:
    post:
      operationId: createPayment
      tags:
        - Payment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: "Платеж создан"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: "Некорректный запрос"
        '500':
          description: "Внутренняя ошибка сервера"
...
components:
  schemas:
    Gym:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор спортзала (UUID)"
        name:
          type: string
          description: "Название спортзала"
        phone:
          type: string
          description: "Телефон спортзала"
        city:
          type: string
          description: "Город, где находится спортзал"
        addres:
          type: string
          description: "Адрес спортзала"
        ischain:
          type: boolean
          description: "Флаг, указывающий, является ли спортзал частью сети"
    PutGym:
      type: object
      required:
        - name
        - phone
        - city
        - addres
        - ischain
      properties:
        name:
          type: string
          description: "Название спортзала"
        phone:
          type: string
          description: "Телефон спортзала"
        city:
          type: string
          description: "Город, где находится спортзал"
        addres:
          type: string
          description: "Адрес спортзала"
        ischain:
          type: boolean
          description: "Флаг, указывающий, является ли спортзал частью сети"
    Amount:
      type: object
      required:
        - value
        - currency
      properties:
        value:
          type: string
          description: "Значение суммы"
        currency:
          type: string
          description: "Валюта"
    Confirmation:
      type: object
      required:
        - type
        - return_url
      properties:
        type:
          type: string
          description: "Тип подтверждения"
        return_url:
          type: string
          description: "URL для возврата после подтверждения"
    PaymentRequest:
      type: object
      required:
        - amount
        - confirmation
        - capture
        - description
      properties:
        amount:
          $ref: '#/components/schemas/Amount'
        confirmation:
          $ref: '#/components/schemas/Confirmation'
        capture:
          type: boolean
          description: "Флаг захвата платежа"
        description:
          type: string
          description: "Описание платежа"
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          description: "Идентификатор платежа"
        status:
          type: string
          description: "Статус платежа"
        paid:
          type: boolean
          description: "Флаг, показывающий, что платеж был успешно завершен"
        amount:
          type: object
          properties:
            value:
              type: string
              description: "Значение суммы"
            currency:
              type: string
              description: "Валюта"
        confirmation:
          type: object
          properties:
            type:
              type: string
              description: "Тип подтверждения"
            return_url:
              type: string
              description: "URL для возврата после подтверждения"
            confirmation_url:
              type: string
              description: "URL для подтверждения платежа"
        created_at:
          type: string
          description: "Дата и время создания платежа"
        description:
          type: string
          description: "Описание платежа"
        metadata:
          type: object
          description: "Дополнительная информация о платеже (метаданные)"
        payment_method:
          type: object
          properties:
            type:
              type: string
              description: "Тип метода оплаты"
            id:
              type: string
              description: "Идентификатор метода оплаты"
            saved:
              type: boolean
              description: "Флаг, указывающий, сохранен ли метод оплаты"
        recipient:
          type: object
          properties:
            account_id:
              type: string
              description: "Идентификатор аккаунта получателя"
            gateway_id:
              type: string
              description: "Идентификатор шлюза получателя"
        refundable:
          type: boolean
          description: "Флаг, указывающий, можно ли вернуть платеж"
        test:
          type: boolean
          description: "Флаг, указывающий, является ли тестовым платежом"
...
    PutEquipment:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: "Название оборудования"
        description:
          type: string
          description: "Описание оборудования"
...
    PutMembershipType:
      type: object
      required:
        - type
        - description
        - price
        - daysduration
      properties:
        type:
          type: string
          description: "Тип членства"
        description:
          type: string
          description: "Описание типа членства"
        price:
          type: string
          description: "Цена членства"
        daysduration:
          type: integer
          description: "Количество дней действия членства"
...
    PutClientMembership:
      type: object
      required:
        - startdate
        - enddate
        - membershiptypeid
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор членства клиента в формате UUID"
        startdate:
          type: string
          format: date
          description: "Дата начала членства в формате YYYY-MM-DD"
        enddate:
          type: string
          format: date
          description: "Дата окончания членства в формате YYYY-MM-DD"
        membershiptypeid:
          type: string
          format: uuid
          description: "Уникальный идентификатор типа членства в формате UUID"
        clientid:
          type: string
          format: uuid
          description: "Уникальный идентификатор клиента в формате UUID"
...
  securitySchemes:
    Bearer:
      type: http
      description: Please enter a valid token
      scheme: bearer
      bearerFormat: JWT